Pal 999 refactored contract KafkaTestConfig for all services
