+++ data-service/src/contract-tests/java/uk/gov/gchq/palisade/contract/data/kafka/KafkaTestConfiguration.java
@@ -14,33 +14,26 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.palisade.contract.topicoffset.kafka;
+package uk.gov.gchq.palisade.contract.data.kafka;
 
 import akka.actor.ActorSystem;
 import akka.stream.Materializer;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.NewTopic;
-import org.apache.kafka.common.serialization.Deserializer;
-import org.apache.kafka.common.serialization.Serializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
 import org.springframework.core.io.ResourceLoader;
-import org.springframework.core.serializer.support.SerializationFailedException;
 import org.testcontainers.containers.KafkaContainer;
 
-import uk.gov.gchq.palisade.service.topicoffset.stream.PropertiesConfigurer;
+import uk.gov.gchq.palisade.service.data.stream.PropertiesConfigurer;
 
-import java.io.IOException;
 import java.util.AbstractMap;
 import java.util.List;
 import java.util.Map;
@@ -64,9 +57,8 @@ public class KafkaTestConfiguration {
     private static final ObjectMapper MAPPER = new ObjectMapper();
 
     private final List<NewTopic> topics = List.of(
-            new NewTopic("masked-resource", 3, (short) 1),
-            new NewTopic("masked-resource-offset", 3, (short) 1),
-            new NewTopic("error", 3, (short) 1));
+            new NewTopic("success", 1, (short) 1),
+            new NewTopic("error", 1, (short) 1));
 
     @Bean
     @ConditionalOnMissingBean
@@ -75,7 +67,7 @@ public class KafkaTestConfiguration {
     }
 
     @Bean
-    KafkaContainer kafkaContainer() throws Exception {
+    KafkaContainer kafkaContainer() throws ExecutionException, InterruptedException {
         final KafkaContainer container = new KafkaContainer("5.5.1")
                 .withReuse(false)
                 .withNetworkMode("host");
@@ -96,7 +88,8 @@ public class KafkaTestConfiguration {
 
     @Bean
     @Primary
-    ActorSystem actorSystem(final PropertiesConfigurer props, final KafkaContainer kafka, final ConfigurableApplicationContext context) {
+    ActorSystem actorSystem(final PropertiesConfigurer props, final KafkaContainer kafka) {
+        LOGGER.info("Starting Kafka with port {}", kafka.getFirstMappedPort());
         return ActorSystem.create("actor-with-overrides", props.toHoconConfig(Stream.concat(
                 props.getAllActiveProperties().entrySet().stream()
                         .filter(kafkaPort -> !kafkaPort.getKey().equals("akka.discovery.config.services.kafka.endpoints[0].port")),
@@ -116,29 +109,5 @@ public class KafkaTestConfiguration {
         String brokers = String.format("%s:%d", "localhost", mappedPort);
         LOGGER.info("brokers: " + brokers);
         return brokers;
-    }
-
-    // Serialiser for upstream test input
-    static class RequestSerializer implements Serializer<JsonNode> {
-        @Override
-        public byte[] serialize(final String s, final JsonNode topicOffsetRequest) {
-            try {
-                return MAPPER.writeValueAsBytes(topicOffsetRequest);
-            } catch (JsonProcessingException e) {
-                throw new SerializationFailedException("Failed to serialize " + topicOffsetRequest.toString(), e);
-            }
-        }
-    }
-
-    // Deserialiser for downstream test output
-    static class ResponseDeserializer implements Deserializer<JsonNode> {
-        @Override
-        public JsonNode deserialize(final String s, final byte[] topicOffsetResponse) {
-            try {
-                return MAPPER.readTree(topicOffsetResponse);
-            } catch (IOException e) {
-                throw new SerializationFailedException("Failed to deserialize " + new String(topicOffsetResponse), e);
-            }
-        }
     }
 }
